Working multi/single drone control

+++ helper/conversions.h
@@ -4,7 +4,7 @@
 #include <mav_trajectory_generation/polynomial_optimization_linear.h>
 #include <mavros_msgs/PositionTarget.h>
 
-mavros_msgs::PositionTarget segmentToPoint(const mav_trajectory_generation::Segment& s, double time);
+mavros_msgs::PositionTarget segmentToPoint(const mav_trajectory_generation::Segment& s, float time);
 
 
 #endif
